#3 separate class to another file
