+++ create/Singleton/src/main_global.py
@@ -1,10 +1,4 @@
-class Config:
-	def __init__(self):
-		self.model_name = "hoge"
-
-	def set_name(self, name):
-		self.model_name = name
-
+from model.config import Config
 
 CONFIG = Config()
