+++ create/Singleton/src/model/config.py
@@ -1,0 +1,6 @@
+class Config:
+	def __init__(self):
+		self.model_name = "hoge"
+
+	def set_name(self, name):
+		self.model_name = name
